backend/requirements.txt
fastapi
uvicorn
pydantic
requests
pillow
